# feat: add comprehensive API performance monitoring and metrics